+++ src/components/feature/index.ts
@@ -1,3 +1,4 @@
 export * from './activity'
 export * from './admin'
+export * from './post'
 export * from './skeleton'
