+++ src/components/feature/post/create-post-form/editor/index.ts
@@ -1,0 +1,1 @@
+export { default as PostContentFieldEditor } from './EditorField'
