+++ src/components/feature/post/create-post-form/index.ts
@@ -1,0 +1,1 @@
+export { CreatePostForm } from './CreatePostForm'
