+++ src/components/feature/post/form-field/index.ts
@@ -1,0 +1,1 @@
+export { PostFormField } from './FormField'
