+++ src/components/feature/post/index.ts
@@ -1,0 +1,2 @@
+export { CreatePostForm } from './create-post-form'
+export { PostFormField } from './form-field'
